% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot.M.R
\name{boot.M}
\alias{boot.M}
\title{Boot mortality}
\usage{
boot.M(data, boot.n)
}
\arguments{
\item{data}{your data}

\item{boot.n}{number of simulations}
}
\description{
Bootstrap Siler mortality
}
\examples{
boot.M()
}
\keyword{bootstrap}

