% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot.Par.R
\name{boot.Par}
\alias{boot.Par}
\title{Bootstraped object}
\usage{
boot.Par(data, boot.n)
}
\arguments{
\item{data}{your data}

\item{boot.n}{number of simulations}
}
\description{
Create an object with bootstrapping parameters
}
\examples{
boot.Par()
}
\keyword{bootstrap}
\keyword{object}

