% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot.Surv.R
\name{boot.Surv}
\alias{boot.Surv}
\title{Bootstrap Siler survivorship}
\usage{
boot.Surv(par, boot.n, rm2 = FALSE)
}
\arguments{
\item{par}{parameters}

\item{boot.n}{number of simulations}

\item{rm2}{remove first two ages. Default = FALSE}
}
\description{
Bootstrap Siler optimized survivorships
}
\examples{
boot.Surv()
}
\keyword{bootstrap}
\keyword{object}
\keyword{survivorship}

