% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot2.M.R
\name{boot2.M}
\alias{boot2.M}
\title{Bootstrap Siler mortality}
\usage{
boot2.M(data, d, rm2 = FALSE)
}
\arguments{
\item{data}{your data}

\item{d}{column of the data}

\item{rm2}{remove first two ages. Default = FALSE}
}
\description{
Bootstrap Siler mortality removing 2 ages
}
\examples{
boot2.M()
}
\keyword{bootstrap}
\keyword{mortality}
\keyword{object}

