% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boot2.Par.R
\name{boot2.Par}
\alias{boot2.Par}
\title{Bootstrapped Siler mortality object}
\usage{
boot2.Par(data, d, rm2 = FALSE)
}
\arguments{
\item{data}{your data}

\item{d}{column of the data}

\item{rm2}{remove first two ages. Default = FALSE}
}
\description{
Create an object with bootstraped parameters removing two ages
}
\examples{
boot2.Par()
}
\keyword{bootstrap}
\keyword{mortality}
\keyword{object}

