% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.ro.R
\name{calc.ro}
\alias{calc.ro}
\title{Calculate rho}
\usage{
calc.ro(A, N.out = FALSE)
}
\arguments{
\item{A}{object}

\item{N.out}{If TRUE output with ro and also numbers. Default = FALSE}
}
\description{
Calculate rho from a Leslie Matrix
}
\examples{
calc.ro()
}
\keyword{leslie}
\keyword{matrix}
\keyword{rho}

