% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coale.R
\name{coale}
\alias{coale}
\title{coale}
\usage{
coale(b1, b4, nMx)
}
\arguments{
\item{b1}{}

\item{b4}{}

\item{nMx}{}
}
\description{
Gives a particular aggregation for life.table(). Original function in demoR_0.4.2
}
\examples{
coale()
}
\keyword{coale}
\keyword{life}
\keyword{table}

