% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort}
\alias{cohort}
\title{cohort}
\usage{
cohort(width12)
}
\arguments{
\item{width12}{}
}
\description{
Gives a particular aggregation for life.table(). Original function in demoR_0.4.2
}
\examples{
cohort()
}
\keyword{cohort}
\keyword{life}
\keyword{table}

