% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dens.prior9.R
\name{dens.prior9}
\alias{dens.prior9}
\title{Density priors}
\usage{
dens.prior9(x, pri.lo = c(0, 0, 0, 0, 0, 0, 15, 0, 0), pri.hi = c(0.15, 1,
  1, 0.5, 0.25, 15, 55, 0.1, 1.25))
}
\arguments{
\item{x}{}

\item{pri.lo}{minimum limits of priors. Default = 0, 0, 0, 0, 0, 0, 15, 0, 0}

\item{pri.hi}{maximum limits of priors Default = 0.15, 1, 1, 0.5, 0.25, 15, 55, 0.1, 1.25}
}
\description{
Modified from dens.prior.R of HPbayes_0.1 package
}
\examples{
dens.prior9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{priors}

