% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw.Siler2.R
\name{draw.Siler2}
\alias{draw.Siler2}
\title{Siler survivorship plot}
\usage{
draw.Siler2(x, deaths = Bot)
}
\arguments{
\item{x}{}

\item{deaths.}{Default = Bot}
}
\description{
Draw a survivorship curve
}
\examples{
draw.Siler2()
}
\keyword{Siler}
\keyword{plot}
\keyword{survivorship}

