% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eigen.analysis.R
\name{eigen.analysis}
\alias{eigen.analysis}
\title{Eigenvalues}
\usage{
eigen.analysis(A)
}
\arguments{
\item{A}{}
}
\description{
Eigenvalues. Original function in demoR_0.4.2
}
\examples{
eigen.analysis()
}
\keyword{eigenvalue}
\keyword{leslie}
\keyword{matrix}

