% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropy.wts.R
\name{entropy.wts}
\alias{entropy.wts}
\title{Entropy}
\usage{
entropy.wts(w)
}
\arguments{
\item{w}{}
}
\description{
Entropy weights. Original script from HPbayes_0.1
}
\examples{
entropy.wts()
}
\keyword{entropy}

