% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expt.upts.R
\name{expt.upts}
\alias{expt.upts}
\title{Exponential updates}
\usage{
expt.upts(w, m)
}
\arguments{
\item{w}{}

\item{m}{}
}
\description{
Exponential updates. Original script from HPbayes_0.1
}
\examples{
expt.upts()
}
\keyword{entropy}

