% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/final.resamp9.R
\name{final.resamp9}
\alias{final.resamp9}
\title{Final resampling}
\usage{
final.resamp9(K, B1, H.new, H.k, log.like, d.keep, prior, h.mu, h.sig, nrisk,
  ndeath, B = 400, theta.dim = 9, age = c(1e-05, 1, seq(5, 100, 5)))
}
\arguments{
\item{K}{}

\item{B1}{}

\item{H.new}{}

\item{H.k}{}

\item{log.like}{}

\item{d.keep}{}

\item{prior}{}

\item{h.mu}{}

\item{h.sig}{}

\item{nrisk}{}

\item{ndeath}{}

\item{B.}{Defalt = 400}

\item{theta.dim.}{Default = 9}

\item{age.}{Default = c(1e-05, 1, seq(5, 100, 5))}
}
\description{
Final resampling. Modified from final.resamp.R of HPbayes_0.1 package
}
\examples{
final.resamp9()
}
\keyword{resampling}

