% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.Siler2.R
\name{fit.Siler2}
\alias{fit.Siler2}
\title{Fit a Siler model}
\usage{
fit.Siler2(x, data)
}
\arguments{
\item{x}{}

\item{data}{your data}
}
\description{
Fit a Siler model
}
\examples{
fit.Siler2()
}
\keyword{Siler}

