% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen.time.R
\name{gen.time}
\alias{gen.time}
\title{Generation time}
\usage{
gen.time(A, peryear = 5)
}
\arguments{
\item{A}{}

\item{x}{}

\item{peryear.}{Default = 5}
}
\description{
Generation time. Original function in demoR_0.4.2
}
\examples{
gen.time()
}
\keyword{generation}
\keyword{leslie}
\keyword{matrix}

