% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hp.bm.imis9.R
\name{hp.bm.imis9}
\alias{hp.bm.imis9}
\title{Heligman Pollard imis}
\usage{
hp.bm.imis9(prior, nrisk, ndeath, K, d = 10, B = 400, age = c(1e-05, 1,
  seq(5, 100, 5)), CI = 95)
}
\arguments{
\item{prior}{}

\item{nrisk}{}

\item{ndeath}{}

\item{K}{}

\item{d.}{Default = 10}

\item{B.}{Default = 400}

\item{age.}{Default c(1e-05, 1, seq(5, 100, 5)), CI = 95)}
}
\description{
Heligman Pollard imis. Modified function from hp.bm.imis.R of HPbayes_0.1
}
\examples{
hp.bm.imis9()
}
\keyword{generation}
\keyword{leslie}
\keyword{matrix}

