% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hp.nqx.bc.R
\name{hp.nqx.bc}
\alias{hp.nqx.bc}
\title{Heligman Pollard mortality rate}
\usage{
hp.nqx.bc(H.out, age = seq(0, 85, 1))
}
\arguments{
\item{H.out}{}

\item{age.}{Default = seq(0, 85, 1)}
}
\description{
Heligman Pollard mortality rate. Modified function from hp.nqx.bc.R of HPbayes_0.1
}
\examples{
hp.nqx.bc()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{mortality}

