% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hp.nqx.nM.R
\name{hp.nqx.nM}
\alias{hp.nqx.nM}
\title{Helligman Pollard mortality}
\usage{
hp.nqx.nM(H.out, age = seq(0, 85, 1))
}
\arguments{
\item{H.out}{}

\item{age.}{Default = seq(0, 85, 1)}
}
\description{
Helligman Pollard mortality. Modified from HPbayes_0.1
}
\examples{
hp.nqx.nM()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{mortality}

