% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hp.nqx9.R
\name{hp.nqx9}
\alias{hp.nqx9}
\title{Helligman Pollard mortality}
\usage{
hp.nqx9(H.out, age = seq(0, 85, 1))
}
\arguments{
\item{H.out}{}

\item{age.}{Default = seq(0, 85, 1)}
}
\description{
Helligman Pollard mortality. Modified from hp.nqx.R of HPbayes_0.1
}
\examples{
hp.nqx9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{mortality}

