% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hpbayes.plot9.R
\name{hpbayes.plot9}
\alias{hpbayes.plot9}
\title{Helligman Pollard plot}
\usage{
hpbayes.plot9(..., nrisk = NULL, ndeath = NULL, age, hpp, yrange = c(0,
  0.8), xrange = c(0, 85), log = FALSE, plotdata = FALSE,
  plotpost = TRUE, data.type = "b", post.type = "l",
  line.col = c("grey", "blue", "red", "dark green"), CI = 95)
}
\arguments{
\item{age}{}

\item{hpp}{}

\item{post.type}{Default = "l"}

\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{yrange.}{Default = c(0, 0.8)}

\item{xrange.}{Default = c(0, 85)}

\item{log.}{Default = FALSE}

\item{plotdata.}{Default = FALSE}

\item{plotpost.}{Default = FALSE}

\item{data.type.}{Default = "b"}

\item{line.col.}{Default = c("grey", "blue", "red", "dark green")}

\item{ci.}{Default = 95}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

