% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HPci.nM.R
\name{HPci.nM}
\alias{HPci.nM}
\title{Helligman Pollard plot}
\usage{
HPci.nM(data, age = age, CI = 95)
}
\arguments{
\item{age}{}

\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

