% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keyfitz.R
\name{keyfitz}
\alias{keyfitz}
\title{keyfitz}
\usage{
keyfitz(b0, nMx)
}
\arguments{
\item{b0}{}

\item{nMx}{}
}
\description{
Gives a particular aggregation for life.table(). Original function in demoR_0.4.2
}
\examples{
keyfitz()
}
\keyword{keyfitz}
\keyword{life}
\keyword{table}

