% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leslie.R
\name{leslie}
\alias{leslie}
\title{Helligman Pollard plot}
\usage{
leslie(lx, mx, L = TRUE, peryear = 5, one.sex = TRUE, SRB = 1.05,
  infant.class = TRUE)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

