% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.Siler.boot.R
\name{life.Siler.boot}
\alias{life.Siler.boot}
\title{Bootstrap of Siler life table}
\usage{
life.Siler.boot(bootM)
}
\arguments{
\item{booM}{}
}
\description{
Bootstrap of Siler life table
}
\examples{
life.Siler.boot()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

