% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.Siler.R
\name{life.Siler}
\alias{life.Siler}
\title{Siler life table}
\usage{
life.Siler(Sqx = DDESiler)
}
\arguments{
\item{Sqx}{Default = DDESiler}
}
\description{
Siler life table
}
\examples{
life.Siler()
}
\keyword{Siler}
\keyword{life}
\keyword{table}

