% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.table.R
\name{life.table}
\alias{life.table}
\title{Helligman Pollard plot}
\usage{
life.table(x, nDx, nKx, b0 = c(0.07, 1.7), b1 = c(0.053, 2.8),
  b4 = c(1.522, 1.518), type = "kf", nxx = 0, iwidth = 5,
  width12 = c(1, 4))
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

