% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/like.resamp9.R
\name{like.resamp9}
\alias{like.resamp9}
\title{Helligman Pollard plot}
\usage{
like.resamp9(K, log.like.0, opt.cov.d, opt.mu.d, d.keep, d = 10,
  theta.dim = 9)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

