% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loop.optim9.R
\name{loop.optim9}
\alias{loop.optim9}
\title{Helligman Pollard plot}
\usage{
loop.optim9(prior, nrisk, ndeath, d = 10, theta.dim = 9, age = c(1e-05, 1,
  seq(5, 100, 5)))
}
\arguments{
\item{age}{}

\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

