% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.bc.R
\name{mod.bc}
\alias{mod.bc}
\title{Helligman Pollard model function}
\usage{
mod.bc(theta, x)
}
\arguments{
\item{theta}{}

\item{x}{}
}
\description{
Helligman Pollard model function. Modified from HPbayes_0.1 package
}
\examples{
mod.bc()
}
\keyword{Heligman}
\keyword{Pollard}

