% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod9p.R
\name{mod9p}
\alias{mod9p}
\title{Helligman Pollard 9 poarameters model function}
\usage{
mod9p(theta, x)
}
\arguments{
\item{theta}{}

\item{x}{}
}
\description{
Helligman Pollard 9 poarameters model function. Modified from mod8p.R of HPbayes_0.1 package
}
\examples{
mod9p()
}
\keyword{Heligman}
\keyword{Pollard}

