% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parBoot.R
\name{parBoot}
\alias{parBoot}
\title{Parameter from bootstraped Siler mortality}
\usage{
parBoot(boot.out, conf = 0.95)
}
\arguments{
\item{boot.out}{}

\item{conf}{Default = 0.95}
}
\description{
Calculate parameters of bootstraped Siler mortality
}
\examples{
parBoot()
}
\keyword{Siler}
\keyword{bootstrap}
\keyword{mortality}
\keyword{parameters}

