% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percBootCI.R
\name{percBootCI}
\alias{percBootCI}
\title{Percentages of bootstraped Siler mortality}
\usage{
percBootCI(boot.out, conf = 0.95)
}
\arguments{
\item{boot.out}{}

\item{conf}{Default = 0.95}
}
\description{
Calculate percentages from bootstraped Siler mortality
}
\examples{
percBootCI()
}
\keyword{Siler}
\keyword{bootstrap}
\keyword{mortality}
\keyword{percentage}

