% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior.form9.R
\name{prior.form9}
\alias{prior.form9}
\title{Helligman Pollard plot}
\usage{
prior.form9(pri.lo = c(0, 0, 0, 0.001, 0, 0, 15, 0, 0), pri.hi = c(0.15, 1,
  1, 0.5, 0.25, 15, 55, 0.1, 1.25), theta.dim = 9)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

