% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior.likewts9.R
\name{prior.likewts9}
\alias{prior.likewts9}
\title{Helligman Pollard 9 parameters priors}
\usage{
prior.likewts9(prior, nrisk, ndeath, theta.dim = 9, age = c(1e-05, 1, seq(5,
  100, 5)))
}
\arguments{
\item{prior}{}

\item{nrisk}{}

\item{ndeath}{}

\item{theta.dim}{Default = 9}

\item{age}{Default = c(1e-05, 1, seq(5, 100, 5))}
}
\description{
Helligman Pollard 9 parameters priors. Modified from prior.likewts.R of HPbayes_0.1 package
}
\examples{
prior.likewts9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{priors}

