% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.leslie.R
\name{project.leslie}
\alias{project.leslie}
\title{Project Leslie matrix}
\usage{
project.leslie(A, no, tmax, pop.sum = FALSE)
}
\arguments{
\item{A}{}

\item{no}{}

\item{tmax}{}

\item{pop.sum}{Default = FALSE}
}
\description{
Project Leslie matrix. Original from demoR_0.4.2
}
\examples{
project.leslie()
}
\keyword{Leslie}
\keyword{matrix}
\keyword{projection}

