% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp.postopt9.R
\name{samp.postopt9}
\alias{samp.postopt9}
\title{Helligman Pollard plot}
\usage{
samp.postopt9(opt.cov.d, opt.mu.d, d.keep, prior, B = 400, B0 = 8000,
  d = 10)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

