% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var.rwts.R
\name{var.rwts}
\alias{var.rwts}
\title{Helligman Pollard variance weights}
\usage{
var.rwts(w)
}
\arguments{
\item{w}{}
}
\description{
Variances weights of Helligman Pollard model. Original script from HPbayes_0.1
}
\examples{
var.rwts()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{weights}

