% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HPci.nM.R
\name{HPci.nM}
\alias{HPci.nM}
\title{Helligman Pollard confidence intervals}
\usage{
HPci.nM(data, age = age, CI = 95)
}
\arguments{
\item{data}{}

\item{age}{Default = age}

\item{CI}{Confidence intervals. Default = 95}
}
\description{
Helligman Pollard confidence intervals. Original function from HPbayes_0.1
}
\examples{
HPci.nM()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{confidence}
\keyword{intervals}

