% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HPci9.R
\name{HPci9}
\alias{HPci9}
\title{Helligman Pollard confidence intervals with 9 parameters}
\usage{
HPci9(data, age = age, CI = 95)
}
\arguments{
\item{data}{}

\item{age}{}

\item{CI}{Confidence intervals. Default = 95}
}
\description{
Helligman Pollard confidence intervals with 9 parameters. Modified from HPci9 of HPbayes_0.1
}
\examples{
HPci9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{confidence}
\keyword{intervals}

