% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leslie.R
\name{leslie}
\alias{leslie}
\title{Leslie matrix}
\usage{
leslie(lx, mx, L = TRUE, peryear = 5, one.sex = TRUE, SRB = 1.05,
  infant.class = TRUE)
}
\arguments{
\item{lx}{}

\item{mx}{}

\item{L}{Default = TRUE}

\item{peryear}{Default = 5}

\item{one.sex}{Default = TRUE}

\item{SRB}{Default = 1.05}

\item{infant.class}{Default = TRUE}
}
\description{
Leslie matrix. Modified from leslie.matrix.R of demoR_0.4.2
}
\examples{
leslie()
}
\keyword{Leslie}
\keyword{matrix}

