% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.Siler.R
\name{life.Siler}
\alias{life.Siler}
\title{Helligman Pollard plot}
\usage{
life.Siler(Sqx = DDESiler)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

