% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life.tab.R
\name{life.tab}
\alias{life.tab}
\title{Life table for Leslie matrix}
\usage{
life.tab(x)
}
\arguments{
\item{x}{}
}
\description{
Life table for Leslie matrix. Modified from life.table.R of demoR_0.4.2
}
\examples{
life.tab()
}
\keyword{Leslie}
\keyword{life}
\keyword{matrix}
\keyword{tale}

