% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/like.resamp9.R
\name{like.resamp9}
\alias{like.resamp9}
\title{Life table resampling}
\usage{
like.resamp9(K, log.like.0, opt.cov.d, opt.mu.d, d.keep, d = 10,
  theta.dim = 9)
}
\arguments{
\item{K}{}

\item{log.like.0}{}

\item{opt.cov.d}{}

\item{opt.mu.d}{}

\item{d.keep}{}

\item{d}{Default = 10}

\item{thete.dim}{Default = 9}
}
\description{
Life table resampling. Modified from like.resamp.R of HPbayes_0.1 package
}
\examples{
like.resamp9()
}
\keyword{life}
\keyword{resampling}
\keyword{table}

