% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ll.binom.R
\name{ll.binom}
\alias{ll.binom}
\title{Density binomial}
\usage{
ll.binom(x, n, p)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Density binomial. Original script from HPbayes_0.1
}
\examples{
ll.binom()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

