% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loop.optim9.R
\name{loop.optim9}
\alias{loop.optim9}
\title{Loop optimization}
\usage{
loop.optim9(prior, nrisk, ndeath, d = 10, theta.dim = 9, age = c(1e-05, 1,
  seq(5, 100, 5)))
}
\arguments{
\item{prior}{}

\item{nrisk}{}

\item{ndeath}{}

\item{d}{Default = 10}

\item{theta.dim}{Default = 9}

\item{age}{Default = c(1e-05, 1, seq(5, 100, 5))}
}
\description{
Loop optimization. modified from loop.optim.R of HPbayes_0.1 package
}
\examples{
loop.optim9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

