% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod.nM.R
\name{mod.nM}
\alias{mod.nM}
\title{Helligman Pollard mortality function}
\usage{
mod.nM(theta, x)
}
\arguments{
\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard mortality function. Modified from HPbayes_0.1
}
\examples{
mod.nM()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{mortality}

