% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parBoot.R
\name{parBoot}
\alias{parBoot}
\title{Helligman Pollard plot}
\usage{
parBoot(boot.out, conf = 0.95)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

