% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.leslie.R
\name{project.leslie}
\alias{project.leslie}
\title{Helligman Pollard plot}
\usage{
project.leslie(A, no, tmax, pop.sum = FALSE)
}
\arguments{
\item{nrisk.}{Default = NULL}

\item{ndeath.}{Default = NULL}

\item{age}{}

\item{hpp}{}
}
\description{
Helligman Pollard plot. Modified from hpbayes.plot.R of HPbayes_0.1
}
\examples{
hpbayes.plot9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{plot}

