% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samp.postopt9.R
\name{samp.postopt9}
\alias{samp.postopt9}
\title{Post-sampling of Helligman Pollard model}
\usage{
samp.postopt9(opt.cov.d, opt.mu.d, d.keep, prior, B = 400, B0 = 8000,
  d = 10)
}
\arguments{
\item{opt.cov.d}{}

\item{opt.mu.d}{}

\item{d.keep}{}

\item{prior}{}

\item{B}{Default = 400}

\item{B0}{Default = 8000}

\item{d}{Default = 10}
}
\description{
Post-sampling of Helligman Pollard model. Modified from samp.postopt.R of HPbayes_0.1 package
}
\examples{
samp.postopt9()
}
\keyword{Heligman}
\keyword{Pollard}
\keyword{post-sampling}

